Fix configuration binding - Update SaleFinder page
