+++ CatalogueScanner.Configuration/IConfigurationSaver.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 
 namespace CatalogueScanner.Configuration
 {
-    public interface IConfigurationService<TOptions>
+    public interface IConfigurationSaver<TOptions>
     {
         Task SaveAsync(TOptions options, CancellationToken cancellationToken = default);
     }
